+++ plugins/reactivity-angular/index.cjs
@@ -1,0 +1,2 @@
+const pkg = require('../../dist/cjs/plugins/reactivity-angular/index.js');
+module.exports = pkg;