+++ plugins/reactivity-preact-signals/index.cjs
@@ -1,0 +1,2 @@
+const pkg = require('../../dist/cjs/plugins/reactivity-preact-signals/index.js');
+module.exports = pkg;