+++ plugins/reactivity-preact-signals/index.ts
@@ -1,0 +1,1 @@
+export * from '../../dist/types/plugins/reactivity-preact-signals/index';